pyproject.toml
[project]
name = "multimodal-ai-agent"
version = "0.1.0"
description = "Local Multimodal AI Agent for document and image management"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "pypdf>=3.0.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "transformers>=4.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]